Added sample tomcat confs.
These are basic tomcat configurations for the
server and users to use the ajp connector and
a user to utilize the manager gui. The user must
be changed in any environment you wish to consider
secure. By default the ssl connection is not enabled
for tomcat. If OpenMRS is the only connection served
from Tomcat this is acceptable as long as port 8080
is only allowed from the localhost.